fix(tagParser): don't break on bad-tags
Rather than throwing an uncaught exception, a bad tag should be added to
the badtags collection so that it can be reported safely, rather than
crashing the doc generation process.

To achieve this, the tagProcessors have been refactored to run on individual
tags rather than on the collection.

Also, this commit removes the deprecated doctrine-tag-parser.

Closes #11