feat(PartialNames): getLink disambiguates docs by area
If there are more than one doc that match a name then the getLink method
fails saying that the name is ambiguous.
Now, if you pass in the current document to getLink then it will first
try to filter out, from the ambiguous docs, those that are in areas
other than the current doc. If this leaves only one doc then that is
returned successfully.