feat(jsdoc package): use `jsdoc` rather than `js` doc-extractor
The `jsdoc` doc-extractor is compatible with the `js` version but provides
additional functionality.  The code following the jsdoc comment block
is provided to the doc as an extra property that can be used by successive
doc-processors.  For instance, to extract the name of the described code
item if no @name tag is provided.