refact(templateFinder): now an object with methods
BREAKING CHANGE:

Before the templateFinder was a function that was generated in a closure
that contained the templatePatterns and templateFolders values.
Now these are private fields set by methods and the overall findTemplate
function has become a method too.

You must now set the templateFolders and templatePatterns in a config
block rather than relying on DI to inject them.