refact(computePathsProcessor): use stringmap and no longer allow default template
BREAKING CHANGE:

This commit removes the default fallback option for path templates. You now
must provide a template for every docType that can appear as doc if you want
them to have a path or outputPath.

If there is no path template defined for a docType then docs of that type
 will quietly not get a path or outputPath.