fix(base/trimIndentation): ensure that lines that start at index zero kill the indentation
Previously we were trying to be clever and guess that lines that had zero indentation were actually
just rolling over from the previous line and so were ignored.

But since we ignore the first line if it has no indentation, we had to rely on the second line having
no indentation for the zero indentation to be set. If the second line had indentation then all future
zero indentation lines were being ignored.

This change now never ignores zero indent lines and so will indent most things properly. If a line is
rolling over then it should be indented to the minimum level already for it to be displayed correctly.

See https://github.com/angular/angular.js/issues/10101