fix(ngdoc/templates): move module installation above description
When learning about an Angular module, it is not obvious sometimes that
a module is external from looking at the docs and trying to learn about it.

Example: ngMessages.

If you never used it before, you might think it comes with Angular and waste
time on debugging why it is not working.

Like on any github readme.md file, I think it might be better to start the
doc with the installation requirements to indicate this is an external module
you need to bower/npm install

Closes #134