feat(inline-tags): allow start-end tags
Such tags have a start and an end marker and the content between the two
tags is passed through to the handler.

For example given the following tag definition:

```js
{
  name: 'block',
  end: 'endblock',
  handler: function(doc, tagName, tagDescription, docs) {
    ...
  }
}
```

and the following rendered content:

```
before {@block param1}content inside block{@endblock} after
```

The handler will be called with the following parameters:

* `tagName` : `'block'`
* `tagDescription : `{ tag: 'param1', content: 'content inside block' }`

You can see that the tag description is now an object, rather than just a string,
containing two properties: `tag` is the text that was extracted from
inside the start tag, `content` is the text that was extracted from between
the start and end tags.

All the renderedContent from the start tag to the end tag, inclusive is
replaced with the return value from the handler.