fix(jsdoc/extract-tags): don't write properties if tag values are `undefined`
Previously undefined values for tags would be written to the doc property,
which meant that you would get an unwanted key on the doc object, even
though the value was `undefined`.

In the case of tagDefs that declared themselves as "multiple", you would
get an unwanted `undefined` item in the array.

BREAKING CHANGE
Tags that had a value of `undefined` are no longer written to the document.
If you relied upon such falsy values forcing keys onto the doc, then change
your tagDefs to return `null` instead.