fix(typescript): correctly extract the type definition from type aliases
When aliasing a class one must create two declarations for a symbol:

```
class OriginalClass {}
const AliasedClass = OriginalClass;
type AliasedClass = OriginalClass;
```

Previously we were looking for the type on the value definition (i.e.
the `const`), and not finding it. Now if the type is not found on the
value definition then we search the other declarations for the first
to have a `type` property.

Related to https://github.com/angular/angular/pull/14523