fix(typescript): include specified namespaces in types
We were stripping all namespace qualifiers on types.

BREAKING CHANGE:

Previously to 0.20.0 the service where you registered namespaces not to strip
was called `ignoreTypeScriptNamespaces` and took an array of regular expressions.

Now this service is called `namespacesToInclude`, which better reflects its
purpose. Also it is now an array of strings, since there was little benefit in it
being a regular expression.