feat(typescript): group function/method overloads together
Exported functions and class methods can have multiple overload declarations
for different call signatures. But there can be at most one real function
implementation.

This changes the docs that are generated for such overloads.
Now there will be a single doc (export or member) for any such overloaded
function, which represents the "real" function.
This doc will then have an `overloads` property that will contain
a collection of docs for each of the other overloads.