feat(typescript): add `realFilePath`/`realProjectRelativePath` to `FileInfo`
The file paths returned by TypeScript may contain symlinks. This is fine most of
the time, since the path points to the correct file, but in certain cases (e.g.
when we want to generate GitHub links based on the project-relative path) it is
necessary to know the real path to the file (without symlinks).

This commit adds two new properties, `realFilePath` and
`realProjectRelativePath`, which are based on the real path to the file.

(Related to angular/angular#18353.)