feat(typescript): change `ClassLikeExportDoc`s to reference ancestor docs
Introduces a change to the API for accessing heritage information and adds
a new processor that wires up the docs by heritage.

BREAKING CHANGE:

Previously, the `doc.implementsClauses` and `doc.extendsClauses` were arrays
of strings, describing the text of the clause.

Now, these properties are of type `HeritageInfo`:

```
class HeritageInfo {
  symbol: Symbol | undefined;
  doc: ClassLikeExportDoc | undefined;
  type: ExpressionWithTypeArguments;
  text: string;
}
```

where the `text` property holds the previous string values.

You need to change your code/templates to access the text of the heritage
clause from:

```
{% for clause in doc.extendsClauses %}{$ clause $}{% endfor %}
```

to:

```
{% for clause in doc.extendsClauses %}{$ clause.text $}{% endfor %}
```