feat(typescript): reference `extands` and `implements` ancestor info
* Introduce a change to the API for accessing heritage information
* Add a new processor that wires up the docs by heritage.

BREAKING CHANGE:

Previously, the `doc.implementsClauses` and `doc.extendsClauses` were arrays
of strings, describing the text of the clause.

Now, these properties are of type `HeritageInfo`:

```
class HeritageInfo {
  symbol: Symbol | undefined;
  doc: ClassLikeExportDoc | undefined;
  type: ExpressionWithTypeArguments;
  text: string;
}
```

where the `text` property holds the value that was previously provided.

You need to change your code/templates to access the text of the heritage
clause from:

```
{% for clause in doc.extendsClauses %}{$ clause $}{% endfor %}
```

to:

```
{% for clause in doc.extendsClauses %}{$ clause.text $}{% endfor %}
```