fix(typescript): do not encode anchors for member API docs
Previously the `anchor` property was URL encoded in the document object.
This constrains the template author.

BREAKING CHANGE:

Anchors to member API docs, whch are accessed via `doc.anchor` are no
longer URL encoded. You should do this in your template if that is required.

Before: `<a href="{$ doc.anchor $}">...</a>`
After: `<a href="{$ doc.anchor | urlencode $}">...</a>`