refactor(typescript): remove `namespacesToInclude` feature
This does not appear to have any usage in current projects
and adds a great deal of complexity to the code base.

BREAKING CHANGE:

The `namespacesToInclude` service has been removed from
the package. Now when rendering types full qualified names are
always rendered and no namespaces are stripped.

If you relied on this then you should implement a processor to
search for these rendered types and strip them.