chore(typescript): use TypeScript v2.7 to parse the TS source files
Closes #263

BREAKING CHANGE:

This new version has better handling for specially named members such
as `__constructor` and `__call`. To support this the new version of
TS introduced a new type `__String` for holding escaped versions of
these strings. This makes is much safer in case a member name
collided, but it also introduces a breaking change to the TypeScript
AST public interface.

This will not affect developers using the `typescript`
package unless they are trying to interact with the TS AST
directly, or they were expected to interact with escaped forms of
special names, such as `__esModule`, which is `___esModule` in
its escaped form.