fix(typescript): empty node description if first leading comment is single line trivia
In case the new `concatMultipleLeadingComments` is set to `false` and the first comment of a TypeScript node is a **single line** non-JSdoc comment, the node content will be empty.

This is because the first single line comment will taken as leading comment but then afterwards filtered because it doesn't start with `/**`.