fix(base): ensure writeFile() actually writes files to disk
The `mkdirp-promise` dependency had its own dependency of
`"mkdirp": "*"`. But recently `mkdirp` did a new breaking change
release that is no longer compatible with `mkdirp-promise`.
This was causing the `writeFile()` service never to resolve its own
promise, which resulted in files not getting written to the disk.

This commit removes the dependency on `mkdirp-promise`
since `mkdirp` now supports returning a promise directly.

Fixes #287